requirements.txt
# Fichier des dépendances Python
biopython>=1.80
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.23.0